Camera: Switch usage flags to 64-bit unsigned
Usage flags should use 64-bit unsigned values.

Bug: 35215313
Test: Camera CTS

+++ services/camera/libcameraservice/device1/CameraHardwareInterface.cpp
@@ -361,7 +361,7 @@ CameraHardwareInterface::setUsage(hardware::graphics::common::V1_0::BufferUsage 
         ALOGE("%s: preview window is null", __FUNCTION__);
         return s;
     }
-    mPreviewUsage = (int) usage;
+    mPreviewUsage = static_cast<uint64_t> (usage);
     int rc = native_window_set_usage(a, mPreviewUsage);
     if (rc == OK) {
         cleanupCirculatingBuffers();
